requirements.txt
# ============================================
# SODA_V3_PHASE1 - All Dependencies
# ============================================

# Dashboard
streamlit==1.41.1
plotly==5.24.1
pandas==2.2.3

# Data Quality Scanner
soda-core-postgres==3.3.16

# Database Driver
psycopg2-binary==2.9.9
